Perf: code-split RowEditor for row peek surface
